Extract method behaviour to a module
