+++ lib/toy/behaviours.rb
@@ -56,5 +56,20 @@ module Toy
         constant_map.keys.map(&:to_sym)
       end
     end
+
+    module Methods
+      def method_map
+        @method_map ||= {}
+      end
+
+      def define_method(name, method)
+        name = name.to_sym
+        method_map[name] = method
+      end
+
+      def instance_methods
+        method_map.keys.map(&:to_sym)
+      end
+    end
   end
 end
